ipmr: Add reference count to MFC entries
Next commits will introduce MFC notifications through the atomic
fib_notification chain, thus allowing modules to be aware of MFC entries.
This mechanism will later be used to allow the mlxsw driver to offload MFC
entries to the Spectrum hardware.

Due to the fact that modules may need to hold a reference to an MFC entry,
allow the modules to prevent the entry from being deleted by adding
reference count. This way, modules registered to the notification chain can
take a reference to the entry while it is being used.

The reference counting is done only on resolved MFC entries currently.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>