mlxsw: spectrum_router: Handle multicast routes notifications
Upon FIB MFC and VIF add and delete notification, call the matching
functions in the multicast routing offload logic.

As the multicast routing offload logic may sleep, the actual calls are done
in a deferred work. To ensure the MFC object is not freed in that interval,
a reference is held to it. In case of a failure, the abort mechanism is
used, which ejects all the routes from the hardware and triggers the
traffic to flow through the kernel.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>