mlxsw: spectrum: Add multicast router traps
Add the traps needed for the hardware multicast router operation. The traps
include the RPF trap which is called when a packet hits a route but arrived
on wrong interface, another general trap that will be used for multicast
routes with trap action, and a trap for PIM control packets.

Signed-off-by: Yotam Gigi <yotamg@mellanox.com>